Use ES5 syntax
Hi, 

We recently had a bug in production due to this library. We use this library on the frontend and we don't transpile the dependencies because it's best practice that npm packages are production ready with ES5 syntax. Many larger libs transpile their code through babel and that could be an option for this library aswell. But because this code is so short I did the easiest thing and just translated the code to ES5 directly.